move max-flow
