+++ test/aoj/grl_6_a.test.cpp
@@ -3,7 +3,7 @@
 #include "src/cpp-template/header/input.hpp"
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
-#include "src/graph/max-flow.hpp"
+#include "src/graph/flow/max-flow.hpp"
 
 #include <iostream>
